HDS-3945 Rename service types file, use types from service in component
