+++ packages/components/src/components/hds/time/index.ts
@@ -7,7 +7,7 @@ import Component from '@glimmer/component';
 import { typeOf } from '@ember/utils';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
-import type { DisplayType } from './types.ts';
+import type { DisplayType } from '../../../services/hds-time-types.ts';
 
 import type TimeService from '../../../services/hds-time';
